server/pyproject.toml
[tool.poetry]
name = "flask-app"
version = "0.1.0"
description = ""
authors = ["andyhewitt <andyhewitt1987@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.2.2"
flask-sqlalchemy = "^3.0.2"
requests = "^2.28.1"
gunicorn = "^20.1.0"
APScheduler = "^3.9.1"
flask-cors = "^3.0.10"
flask-migrate = "^4.0.1"
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
